+++ chatbot-20220408/alibabacloud_chatbot20220408/models.py
@@ -364,9 +364,13 @@ class BeginSessionRequest(TeaModel):
         self,
         agent_key: str = None,
         instance_id: str = None,
+        sand_box: bool = None,
+        vendor_param: str = None,
     ):
         self.agent_key = agent_key
         self.instance_id = instance_id
+        self.sand_box = sand_box
+        self.vendor_param = vendor_param
 
     def validate(self):
         pass
@@ -381,6 +385,10 @@ class BeginSessionRequest(TeaModel):
             result['AgentKey'] = self.agent_key
         if self.instance_id is not None:
             result['InstanceId'] = self.instance_id
+        if self.sand_box is not None:
+            result['SandBox'] = self.sand_box
+        if self.vendor_param is not None:
+            result['VendorParam'] = self.vendor_param
         return result
 
     def from_map(self, m: dict = None):
@@ -389,6 +397,10 @@ class BeginSessionRequest(TeaModel):
             self.agent_key = m.get('AgentKey')
         if m.get('InstanceId') is not None:
             self.instance_id = m.get('InstanceId')
+        if m.get('SandBox') is not None:
+            self.sand_box = m.get('SandBox')
+        if m.get('VendorParam') is not None:
+            self.vendor_param = m.get('VendorParam')
         return self